requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.1
aiogram==3.14.0
APScheduler==3.10.4

pytest==8.3.3
pytest-cov==6.0.0
requests-mock==1.12.1
pytest-asyncio==0.24.0

flake8==7.1.1
black==24.10.0
isort==5.13.2